+++ xknx/io/request_response/session.py
@@ -12,8 +12,10 @@ if TYPE_CHECKING:
     from xknx.io.transport import KNXIPTransport
 
 
-class Session(RequestResponse):
+class Session(RequestResponse[SessionResponse]):
     """Class to send a SessionRequest and wait for SessionResponse."""
+
+    __slots__ = ("ecdh_client_public_key", "response")
 
     def __init__(
         self,
@@ -22,11 +24,10 @@ class Session(RequestResponse):
     ) -> None:
         """Initialize Session class."""
         # TODO: increase timeout to timeoutAuthentication: 10sec ?
-        super().__init__(transport, SessionResponse)
+        super().__init__(transport)
         self.ecdh_client_public_key = ecdh_client_public_key
-        # TODO: make RequestResponse generic for response class
-        # maybe replace self.success with self.response None check
-        # remove on_success_hook in favour of using knxipframe.body directly
+        # TODO: maybe replace self.success with self.response None check
+        # and remove on_success_hook in favour of using self.response directly
         self.response: SessionResponse | None = None
 
     def create_knxipframe(self) -> KNXIPFrame:
@@ -35,7 +36,6 @@ class Session(RequestResponse):
             SessionRequest(ecdh_client_public_key=self.ecdh_client_public_key)
         )
 
-    def on_success_hook(self, knxipframe: KNXIPFrame) -> None:
+    def on_success_hook(self, response: SessionResponse) -> None:
         """Set communication channel and identifier after having received a valid answer."""
-        assert isinstance(knxipframe.body, SessionResponse)
-        self.response = knxipframe.body
+        self.response = response