+++ xknx/io/request_response/tunnelling.py
@@ -12,8 +12,10 @@ if TYPE_CHECKING:
     from xknx.io.transport import UDPTransport
 
 
-class Tunnelling(RequestResponse):
+class Tunnelling(RequestResponse[TunnellingAck]):
     """Class to send TunnelingRequest and wait for TunnelingACK (UDP only)."""
+
+    __slots__ = ("data_endpoint_addr", "tunnelling_request")
 
     transport: UDPTransport
 
@@ -26,7 +28,7 @@ class Tunnelling(RequestResponse):
         """Initialize Tunnelling class."""
         self.data_endpoint_addr = data_endpoint
         self.tunnelling_request = tunnelling_request
-        super().__init__(transport, TunnellingAck)
+        super().__init__(transport)
 
     async def send_request(self) -> None:
         """Build knxipframe (within derived class) and send via UDP."""